Cherry-pick migrate highlight
